+++ src/Servers/IIS/IIS/test/IIS.Tests/MaxRequestBodySizeTests.cs
@@ -11,6 +11,7 @@ using Microsoft.AspNetCore.Server.IIS.FunctionalTests;
 using Microsoft.AspNetCore.Testing;
 using Microsoft.Extensions.Logging.Testing;
 using Xunit;
+using BadHttpRequestException = Microsoft.AspNetCore.Http.BadHttpRequestException;
 
 namespace IIS.Tests
 {
@@ -23,9 +24,8 @@ namespace IIS.Tests
         {
             var globalMaxRequestBodySize = 0x100000000;
 
-#pragma warning disable CS0618 // Type or member is obsolete
             BadHttpRequestException exception = null;
-#pragma warning restore CS0618 // Type or member is obsolete
+
             using (var testServer = await TestServer.Create(
                 async ctx =>
                 {
@@ -33,9 +33,7 @@ namespace IIS.Tests
                     {
                         await ctx.Request.Body.ReadAsync(new byte[2000]);
                     }
-#pragma warning disable CS0618 // Type or member is obsolete
                     catch (BadHttpRequestException ex)
-#pragma warning restore CS0618 // Type or member is obsolete
                     {
                         exception = ex;
                         throw ex;
@@ -63,9 +61,8 @@ namespace IIS.Tests
             var maxRequestSize = 0x10000;
             var perRequestMaxRequestBodySize = 0x100;
 
-#pragma warning disable CS0618 // Type or member is obsolete
             BadHttpRequestException exception = null;
-#pragma warning restore CS0618 // Type or member is obsolete
+
             using (var testServer = await TestServer.Create(
                 async ctx =>
                 {
@@ -77,9 +74,8 @@ namespace IIS.Tests
 
                         await ctx.Request.Body.ReadAsync(new byte[2000]);
                     }
-#pragma warning disable CS0618 // Type or member is obsolete
+
                     catch (BadHttpRequestException ex)
-#pragma warning restore CS0618 // Type or member is obsolete
                     {
                         exception = ex;
                         throw ex;
@@ -273,9 +269,8 @@ namespace IIS.Tests
         {
             var maxRequestSize = 0x1000;
 
-#pragma warning disable CS0618 // Type or member is obsolete
             BadHttpRequestException exception = null;
-#pragma warning restore CS0618 // Type or member is obsolete
+
             using (var testServer = await TestServer.Create(
                 async ctx =>
                 {
@@ -286,9 +281,7 @@ namespace IIS.Tests
                             var num = await ctx.Request.Body.ReadAsync(new byte[2000]);
                         }
                     }
-#pragma warning disable CS0618 // Type or member is obsolete
                     catch (BadHttpRequestException ex)
-#pragma warning restore CS0618 // Type or member is obsolete
                     {
                         exception = ex;
                         throw ex;
@@ -317,20 +310,17 @@ namespace IIS.Tests
         [ConditionalFact]
         public async Task EveryReadFailsWhenContentLengthHeaderExceedsGlobalLimit()
         {
-#pragma warning disable CS0618 // Type or member is obsolete
             BadHttpRequestException requestRejectedEx1 = null;
             BadHttpRequestException requestRejectedEx2 = null;
-#pragma warning restore CS0618 // Type or member is obsolete
+
             using (var testServer = await TestServer.Create(
                 async ctx =>
                 {
                     var buffer = new byte[1];
-#pragma warning disable CS0618 // Type or member is obsolete
                     requestRejectedEx1 = await Assert.ThrowsAsync<BadHttpRequestException>(
                         async () => await ctx.Request.Body.ReadAsync(buffer, 0, 1));
                     requestRejectedEx2 = await Assert.ThrowsAsync<BadHttpRequestException>(
                         async () => await ctx.Request.Body.ReadAsync(buffer, 0, 1));
-#pragma warning restore CS0618 // Type or member is obsolete
                     throw requestRejectedEx2;
                 }, LoggerFactory, new IISServerOptions { MaxRequestBodySize = 0 }))
             {
