+++ src/Servers/Kestrel/Core/src/Internal/Http/HttpParser.cs
@@ -225,9 +225,8 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 
                         Debug.Assert(readAhead == 0 || readAhead == 2);
                         // Headers don't end in CRLF line.
-#pragma warning disable CS0618 // Type or member is obsolete
-                        BadHttpRequestException.Throw(RequestRejectionReason.InvalidRequestHeadersNoCRLF);
-#pragma warning restore CS0618 // Type or member is obsolete
+
+                        KestrelBadHttpRequestException.Throw(RequestRejectionReason.InvalidRequestHeadersNoCRLF);
                     }
 
                     var length = 0;
@@ -454,12 +453,11 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
         [MethodImpl(MethodImplOptions.NoInlining)]
 #pragma warning disable CS0618 // Type or member is obsolete
         private unsafe BadHttpRequestException GetInvalidRequestException(RequestRejectionReason reason, byte* detail, int length)
-
-            => BadHttpRequestException.GetException(
+#pragma warning restore CS0618 // Type or member is obsolete
+            => KestrelBadHttpRequestException.GetException(
                 reason,
                 _showErrorDetails
                     ? new Span<byte>(detail, length).GetAsciiStringEscaped(Constants.MaxExceptionDetailSize)
                     : string.Empty);
-#pragma warning restore CS0618 // Type or member is obsolete
     }
 }
