+++ src/Servers/Kestrel/samples/SampleApp/Startup.cs
@@ -17,7 +17,6 @@ using Microsoft.AspNetCore.Server.Kestrel.Https;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using BadHttpRequestException = Microsoft.AspNetCore.Server.Kestrel.Core.BadHttpRequestException;
 
 namespace SampleApp
 {
@@ -37,9 +36,7 @@ namespace SampleApp
                 {
                     await next.Invoke();
                 }
-#pragma warning disable CS0618 // Type or member is obsolete
-                catch (BadHttpRequestException ex) when (ex.StatusCode == StatusCodes.Status413RequestEntityTooLarge) { }
-#pragma warning restore CS0618 // Type or member is obsolete
+                catch (Microsoft.AspNetCore.Http.BadHttpRequestException ex) when (ex.StatusCode == StatusCodes.Status413RequestEntityTooLarge) { }
             });
 
             app.Run(async context =>
