+++ src/Servers/Kestrel/shared/test/CompositeKestrelTrace.cs
@@ -104,9 +104,7 @@ namespace Microsoft.AspNetCore.Testing
             _trace2.NotAllConnectionsClosedGracefully();
         }
 
-#pragma warning disable CS0618 // Type or member is obsolete
-        public void ConnectionBadRequest(string connectionId, BadHttpRequestException ex)
-#pragma warning restore CS0618 // Type or member is obsolete
+        public void ConnectionBadRequest(string connectionId, Microsoft.AspNetCore.Http.BadHttpRequestException ex)
         {
             _trace1.ConnectionBadRequest(connectionId, ex);
             _trace2.ConnectionBadRequest(connectionId, ex);
