+++ src/Servers/Kestrel/test/InMemory.FunctionalTests/Http2/Http2TimeoutTests.cs
@@ -144,9 +144,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
 
             _mockTimeoutHandler.Verify(h => h.OnTimeout(TimeoutReason.RequestHeaders), Times.Once);
 
-#pragma warning disable CS0618 // Type or member is obsolete
-            await WaitForConnectionErrorAsync<BadHttpRequestException>(
-#pragma warning restore CS0618 // Type or member is obsolete
+            await WaitForConnectionErrorAsync<Microsoft.AspNetCore.Http.BadHttpRequestException>(
                 ignoreNonGoAwayFrames: false,
                 expectedLastStreamId: int.MaxValue,
                 Http2ErrorCode.INTERNAL_ERROR,
